+++ CHANGELOG.md
@@ -14,6 +14,10 @@ This release adds support for React Native 0.77 to `@bugsnag/react-native`
 - (plugin-angular) Added a null check so BugsnagErrorHandler fails silently when misconfigured [#2295](https://github.com/bugsnag/bugsnag-js/pull/2295)
 - (react-native) Fix no such module error when importing Bugsnag from Swift [#2335](https://github.com/bugsnag/bugsnag-js/pull/2335)
 
+### Changed
+
+- (react-native-cli) Remove the installation of the BugSnag Android Gradle Plugin [#2346](https://github.com/bugsnag/bugsnag-js/pull/2346)
+
 ## [8.2.0] - 2025-01-27
 
 ### Added