+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [8.3.0] - 2025-03-04
+
+This release adds support for React Native 0.77 and 0.78 to `@bugsnag/react-native`
+
+### Added
+
+- (react-native-cli) Add support for `AppDelegate.swift` files to insert command [#2319](https://github.com/bugsnag/bugsnag-js/pull/2319)
+- (plugin-react-navigation) Add support for React Navigation v7.x [#2335](https://github.com/bugsnag/bugsnag-js/pull/2335)
+
+### Fixed
+
+- (plugin-angular) Added a null check so BugsnagErrorHandler fails silently when misconfigured [#2295](https://github.com/bugsnag/bugsnag-js/pull/2295)
+- (react-native) Fix no such module error when importing Bugsnag from Swift [#2335](https://github.com/bugsnag/bugsnag-js/pull/2335)
+- (react-native) Fix turbo module check in React Native 0.77 [#2341](https://github.com/bugsnag/bugsnag-js/pull/2341)
+
+### Changed
+
+- (react-native-cli) Remove the installation of the BugSnag Android Gradle Plugin [#2346](https://github.com/bugsnag/bugsnag-js/pull/2346)
+
 ## [8.2.0] - 2025-01-27
 
 ### Added